Moved formatDuration to core/v3
