+++ packages/core/package.json
@@ -73,6 +73,7 @@
     "@opentelemetry/sdk-trace-base": "^1.21.0",
     "@opentelemetry/sdk-trace-node": "^1.21.0",
     "@opentelemetry/semantic-conventions": "^1.21.0",
+    "humanize-duration": "^3.27.3",
     "ulidx": "^2.2.1",
     "zod": "3.22.3",
     "zod-error": "1.5.0"
@@ -80,6 +81,7 @@
   "devDependencies": {
     "@trigger.dev/tsconfig": "workspace:*",
     "@trigger.dev/tsup": "workspace:*",
+    "@types/humanize-duration": "^3.27.1",
     "@types/jest": "^29.5.3",
     "@types/node": "16",
     "jest": "^29.6.2",
