+++ packages/core/src/v3/index.ts
@@ -8,6 +8,12 @@ export * from "./runtime-api";
 export * from "./logger-api";
 export { SemanticInternalAttributes } from "./semanticInternalAttributes";
 export { iconStringForSeverity } from "./icons";
+export {
+  formatDuration,
+  formatDurationMilliseconds,
+  formatDurationNanoseconds,
+  formatDurationInDays,
+} from "./utils/durations";
 
 export function parseTriggerTaskRequestBody(body: unknown) {
   return TriggerTaskRequestBody.safeParse(body);
